Move options code to a testable helper
